OMRReaderServer/requirements.txt
Django==1.11.29
djangorestframework==3.9.4
gunicorn==19.10.0
pytz==2020.4